<fix> 댓글 create 오류 수정

+++ src/main/java/com/hanghae/mungnayng/service/CommentService.java
@@ -3,7 +3,6 @@ package com.hanghae.mungnayng.service;
 import com.hanghae.mungnayng.domain.comment.Comment;
 import com.hanghae.mungnayng.domain.comment.dto.CommentRequestDto;
 import com.hanghae.mungnayng.domain.comment.dto.CommentResponseDto;
-import com.hanghae.mungnayng.domain.item.Item;
 import com.hanghae.mungnayng.domain.member.Member;
 import com.hanghae.mungnayng.repository.CommentRepository;
 import com.hanghae.mungnayng.repository.ItemRepository;
@@ -14,41 +13,40 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
-@Service
-//@AllArgsConstructor
 @RequiredArgsConstructor
+@Service
 public class CommentService {
     private final CommentRepository commentRepository;
     private final ItemRepository itemRepository;
 
-//댓글 생성
+    //댓글 생성
     @Transactional
-public CommentResponseDto createComment(Member member,Long itemId ,CommentRequestDto requestDto) {
-    Item item = itemRepository.findById(itemId)
-            .orElseThrow(() -> new IllegalArgumentException( "게시글이 없습니다."));
+    public CommentResponseDto createComment(Member member, Long itemId, CommentRequestDto requestDto) {
         Comment comment = Comment.builder()
-                .item(item)
+                .nickname(member.getNickname())
                 .content(requestDto.getContent())
                 .build();
         commentRepository.save(comment);
+        if (!itemId.equals(comment.getItem().getId()))
+            throw new IllegalArgumentException("해당 게시글이 없습니다");
         return CommentResponseDto.All(comment);
     }
-//*댓글 조회(댓글 카운트 값, 게시물 아이디 값 아직 안 넣음)//
+
+    //*댓글 조회(댓글 카운트 값, 게시물 아이디 값 아직 안 넣음)//
     @Transactional
     public List<CommentResponseDto> getComment(Long itemId) {
         List<Comment> allByCommentList = commentRepository.findAllById(itemId);
-            return allByCommentList.stream()
-                    .map(CommentResponseDto :: All)
-                    .collect(Collectors.toList());
+        return allByCommentList.stream()
+                .map(CommentResponseDto::All)
+                .collect(Collectors.toList());
     }
 //*댓글 수정//
     @Transactional
-    public CommentResponseDto updateComment(Member member,Long itemId, Long commentId,  CommentRequestDto requestDto) {
-        Item item = itemRepository.findById(itemId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다"));
+    public CommentResponseDto updateComment(Member member, Long itemId, Long commentId, CommentRequestDto requestDto) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 댓글이 없습니다."));
+        if (!itemId.equals(comment.getItem().getId()))
+            throw new IllegalArgumentException("해당 게시글이 없습니다");
         if (!member.getNickname().equals(comment.getNickname())) {
             throw new IllegalArgumentException("작성자 닉네임이 일치하지 않습니다.");
         }
@@ -58,18 +56,18 @@ public CommentResponseDto createComment(Member member,Long itemId ,CommentReques
 
     //*댓글 삭제 미완//
     @Transactional
-        public void deleteComment(Member member, Long itemId, Long commentId) {
-        Item item = itemRepository.findById(itemId)
-                .orElseThrow(()-> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+    public void deleteComment(Member member, Long itemId, Long commentId) {
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(()-> new IllegalArgumentException("해당 코멘트가 존재하지 않습니다."));
-        if(!member.getNickname().equals(comment.getNickname())) {
+                .orElseThrow(() -> new IllegalArgumentException("해당 코멘트가 존재하지 않습니다."));
+        if (!itemId.equals(comment.getItem().getId()))
+            throw new IllegalArgumentException("해당 게시글이 없습니다");
+        if (!member.getNickname().equals(comment.getNickname())) {
             throw new IllegalArgumentException("작성자 닉네임이 일치하지 않습니다.");
         }
         commentRepository.delete(comment);
     }
 
-    }
+}
 
 
 //    @Transactional(readOnly = true)
